feat: enhance membership deletion logic and revalidation

+++ src/features/membership/actions/delete-membership.ts
@@ -4,6 +4,9 @@ import { toActionState } from '@/components/form/utils/to-action-state';
 import { getAuthOrRedirect } from '@/features/auth/queries/get-auth-or-redirect';
 import { prisma } from '@/lib/prisma';
 import { getMemberships } from '../queries/get-memberships';
+import { revalidatePath } from 'next/cache';
+import { setCookieByKey } from '@/actions/cookies';
+import { membershipsPathFor } from '@/paths';
 
 export const deleteMembership = async ({
   userId,
@@ -12,7 +15,7 @@ export const deleteMembership = async ({
   userId: string;
   organizationId: string;
 }) => {
-  await getAuthOrRedirect();
+  const { user } = await getAuthOrRedirect();
 
   const memberships = await getMemberships(organizationId);
 
@@ -25,6 +28,46 @@ export const deleteMembership = async ({
     );
   }
 
+   // Check if membership exists
+   const targetMembership = (memberships ?? []).find(
+    (membership) => membership.userId === userId
+  );
+
+  if (!targetMembership) {
+    return toActionState("ERROR", "Membership not found");
+  }
+
+  // Check if user is deleting last admin
+  const adminMemberships = (memberships ?? []).filter(
+    (membership) => membership.membershipRole === "ADMIN"
+  );
+
+  const removesAdmin = targetMembership.membershipRole === "ADMIN";
+  const isLastAdmin = adminMemberships.length <= 1;
+
+  if (removesAdmin && isLastAdmin) {
+    return toActionState(
+      "ERROR",
+      "You cannot delete the last admin of an organization"
+    );
+  }
+
+  // Check if user is deleting last admin
+  const myMembership = (memberships ?? []).find(
+    (membership) => membership.userId === user?.id
+  );
+
+  const isMyself = user.id === userId;
+  const isAdmin = myMembership?.membershipRole === "ADMIN";
+
+  if (!isMyself && !isAdmin) {
+    return toActionState(
+      "ERROR",
+      "You can only delete memberships as an admin"
+    );
+  }
+
+
   await prisma.membership.delete({
     where: {
       membershipId: {
@@ -34,5 +77,12 @@ export const deleteMembership = async ({
     },
   });
 
-  return toActionState('SUCCESS', 'The membership has been deleted');
+  revalidatePath(membershipsPathFor(organizationId));
+
+  await setCookieByKey(
+    "toast",
+    isMyself
+      ? "You have left the organization"
+      : "The membership has been deleted"
+  );
 };
